+++ src/ext/diff/split_diff_view.js
@@ -1,21 +1,13 @@
 "use strict";
 
-var LineWidgets = require("../../line_widgets").LineWidgets;
-
 var BaseDiffView = require("./base_diff_view").BaseDiffView;
 var config = require("../../config");
 
-class DiffView extends BaseDiffView {
+class SplitDiffView extends BaseDiffView {
     /**
      * Constructs a new side by side DiffView instance.
      *
-     * @param {Object} [diffModel] - The model for the diff view.
-     * @param {import("../../editor").Editor} [diffModel.editorA] - The editor for the original view.
-     * @param {import("../../editor").Editor} [diffModel.editorB] - The editor for the edited view.
-     * @param {import("../../edit_session").EditSession} [diffModel.sessionA] - The edit session for the original view.
-     * @param {import("../../edit_session").EditSession} [diffModel.sessionB] - The edit session for the edited view.
-     * @param {string} [diffModel.valueA] - The original content.
-     * @param {string} [diffModel.valueB] - The modified content.
+     * @param {import("../diff").DiffModel} [diffModel] - The model for the diff view.
      */
     constructor(diffModel) {
         diffModel = diffModel || {};
@@ -220,4 +212,4 @@ class DiffView extends BaseDiffView {
 }
 
 
-exports.DiffView = DiffView;
+exports.SplitDiffView = SplitDiffView;